backend/requirements.txt
asgiref==3.3.0
attrs==20.3.0
boto3==1.16.13
botocore==1.19.13
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
defusedxml==0.7.0rc1
Django==3.1.3
django-3-jet @ https://github.com/Barukimang/django-jet/archive/dev.zip
django-cors-headers==3.5.0
django-dry-rest-permissions==1.1.0
django-environ==0.4.5
django-filter==2.4.0
django-storages==1.10.1
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-gis==0.16
djangorestframework-simplejwt==4.5.0
djoser==2.1.0
drf-spectacular==0.11.0
feedparser==6.0.2
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
MarkupSafe==1.1.1
oauthlib==3.1.0
Pillow==8.0.1
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.4
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
sgmllib3k==1.0.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.25.11
uWSGI==2.0.19.1
